requirements.txt
africastalking==1.2.7
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.0.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
django4-background-tasks==1.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.0
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
schema==0.7.7
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.3
